+++ src/processor.cpp
@@ -10,9 +10,9 @@ float Processor::Utilization() {
   // Get CPU times from the "cpu" line
   std::vector<std::string> cpuUtilStr = LinuxParser::CpuUtilization();
 
-  // Convert CPU times from string to unsigned int
-  unsigned int value;
-  std::vector<unsigned int> cpuUtil;
+  // Convert CPU times from string to long
+  long value;
+  std::vector<long> cpuUtil;
   for (std::size_t i = 1; i < cpuUtilStr.size(); ++i) {
     std::istringstream(cpuUtilStr[i]) >> value;
     cpuUtil.push_back(value);
@@ -22,13 +22,13 @@ float Processor::Utilization() {
   // - 0:user, 1:nice, 2:system, 3:idle, 4:iowait, 5:irq,
   // - 6:softirq, 7:steal, 8:guest, 9:guest_nice
   // idle + iowait
-  unsigned int idle = cpuUtil[3] + cpuUtil[4];
+  long idle = cpuUtil[3] + cpuUtil[4];
   // user + nice + system + irq + softirq + steal
-  unsigned int nonIdle = cpuUtil[0] + cpuUtil[1] + cpuUtil[2] + cpuUtil[5] +
-                         cpuUtil[6] + cpuUtil[7];
-  unsigned int total = idle + nonIdle;
-  unsigned int d_total = total - this->p_total;
-  unsigned int d_idle = idle - this->p_idle;
+  long nonIdle = cpuUtil[0] + cpuUtil[1] + cpuUtil[2] + cpuUtil[5] +
+                 cpuUtil[6] + cpuUtil[7];
+  long total = idle + nonIdle;
+  long d_total = total - this->p_total;
+  long d_idle = idle - this->p_idle;
   this->p_total = total;
   this->p_idle = idle;
